requirements/common.txt
alembic==0.7.6
flasgger-TheodoFork==0.5.13
Flask-Cors==3.0.6
Flask-Migrate==1.8.0
Flask-RESTful==0.3.7
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask==1.0.2
Mako==1.0.1
MarkupSafe==0.23
mysqlclient==1.3.7
raven==5.8.1
requests==2.20.1
SQLAlchemy==1.1.3
Werkzeug==0.14.1
